Sortable fix for Issue #22
Seems this was simpler than expected, previously I was simply comparing
the first character of the string which as expected produced incorrect
results. I have now changed this to use the entire string for
comparisons which fixes the sorting issues. Must just keep an eye on
this in case of performance issues but it seems fine.