Group Header detail row support Issue #23
Changed the way the default createDetail function works in order to
accomodate the new Group Header features.
In order for this to be accomplished the cell data passed to the
createDetail function now contains 2 additional properties, **group**
(The idetifier used in the data-group attribute) and **groupName** (The
actual text value of the group).