Changes and fixes for HTMLColumn sorting
Changed the default sortValue for an HTML column cell to be the raw HTML
string of the cell's contents. There was a bug where it was using the
actual DOM elements themselves for sorting. This can be changed using
the HTML column specific attribute, "data-sort-use", to instead use the
text of the cell's content by setting the value to "text".