#444 - Added in some new events to the plugin initialization and refactored the constructor logic into a separate method so it can be easily overridden.
postinit.ft.table - This is raised directly after all components have been initialized but before the table is drawn. Calling event.preventDefault() will stop the drawing of the table.
ready.ft.table - This is raised after all components have been initialized and be default drawn. If the postinit.ft.table event has it's default prevented then the table is not drawn when this is event is raised. Calling event.preventDefault will stop the executing of the ready callback.