Merge supplied 'data' to 'this.value' instead of 'this.value = data'
In some instances, the user might supply values that does not contain all the key:value that are present in the original row's data.
When running row.val(data, redraw) or ft.rows.update(indexOrRow, data, redraw), we will then lose original data if the key:value is not supplied in the new 'data'. Doing a merge of the objects will prevent this.