Added a specific class for footable's default search filter
Added search-filter class to the the default search filter element - so we can target it with css or js - e.g. to hide it in a scenario where we have other custom filtering elements inside `.footable-filtering` so we can't hide the whole `.footable-filtering` table row as mentioned in the examples.